Move Hypermedia types to context
