+++ src/context.rs
@@ -15,7 +15,6 @@ use HttpVersion;
 use Method;
 use header::Headers;
 use log::Log;
-use router::Hypermedia;
 
 use Global;
 
@@ -108,4 +107,39 @@ impl<'a, 'b> ExtQueryBody for BodyReader<'a, 'b> {
         try!(self.read_to_end(&mut buf));
         Ok(utils::parse_parameters(&buf))
     }
+}
+
+///Hypermedia connected to an API endpoint.
+pub struct Hypermedia<'a> {
+    ///Forward links from the current endpoint to other endpoints.
+    pub links: Vec<Link<'a>>
+}
+
+impl<'a> Hypermedia<'a> {
+    pub fn new() -> Hypermedia<'a> {
+        Hypermedia {
+            links: vec![]
+        }
+    }
+}
+
+///A hyperlink.
+#[derive(PartialEq, Eq, Debug)]
+pub struct Link<'a> {
+    ///The HTTP method for which an endpoint is available. It can be left
+    ///unspecified if the method doesn't matter.
+    pub method: Option<Method>,
+    ///A relative path from the current location.
+    pub path: Vec<LinkSegment<'a>>
+}
+
+///A segment of a hyperlink path.
+#[derive(PartialEq, Eq, Debug)]
+pub enum LinkSegment<'a> {
+    ///A static part of a path.
+    Static(&'a str),
+    ///A dynamic part of a path. Can be substituted with anything.
+    Variable(&'a str),
+    ///A recursive wildcard. Will recursively match anything.
+    RecursiveWildcard
 }
